pyproject.toml
[tool.poetry]
name = "hybrid-rag-app"
version = "2.0.0"
description = "Advanced Hybrid RAG Application with sparse + dense embeddings, metadata enrichment, and local LLM"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}, {include = "components"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
haystack-ai = ">=2.8.0"
ollama-haystack = ">=0.0.8"
qdrant-haystack = ">=4.0.0"
fastembed-haystack = ">=1.0.0"
sentence-transformers = ">=3.0.0"
pypdf = ">=3.0.0"
python-dotenv = ">=1.0.0"
trafilatura = ">=1.12.0"
tqdm = ">=4.66.0"
rich = ">=13.7.0"
accelerate = ">=0.20.0"
transformers = ">=4.30.0"
langfuse = ">=2.0.0"
qdrant-client = "^1.16.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"

[tool.poetry.group.gpu]
optional = true

[tool.poetry.group.gpu.dependencies]
torch = ">=2.0.0"

[tool.poetry.scripts]
index-docs = "index_documents:main"
query = "query_app:main"
investigate = "investigate_documents:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
